app/api/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.7
GitPython==3.1.43
redis==5.0.8
rq==1.16.2
PyYAML==6.0.2
markdown-it-py==3.0.0
mdurl==0.1.2
Jinja2==3.1.4
requests==2.32.3
orjson==3.10.7
aiofiles==24.1.0
tenacity==9.0.0
structlog==24.4.0
pytest==8.3.3
ruff==0.6.8
black==24.8.0
mypy==1.11.2

